reduce test make targets, organize test file structure
