+++ .ibm/pipelines/windows-test-script.ps1
@@ -67,15 +67,9 @@ function Run-Test {
     Shout "Create Binary"
     make install 
     Shout "Running test"
-    make test-integration-devfile   | tee -a  C:\Users\Administrator.ANSIBLE-TEST-VS\AppData\Local\Temp\$LOGFILE
-    Check-ExitCode $LASTEXITCODE
     make test-integration           | tee -a  C:\Users\Administrator.ANSIBLE-TEST-VS\AppData\Local\Temp\$LOGFILE
     Check-ExitCode $LASTEXITCODE
-    make test-cmd-login-logout      | tee -a  C:\Users\Administrator.ANSIBLE-TEST-VS\AppData\Local\Temp\$LOGFILE
-    Check-ExitCode $LASTEXITCODE
-    make test-cmd-project           | tee -a  C:\Users\Administrator.ANSIBLE-TEST-VS\AppData\Local\Temp\$LOGFILE
-    Check-ExitCode $LASTEXITCODE
-    make test-e2e-devfile           | tee -a  C:\Users\Administrator.ANSIBLE-TEST-VS\AppData\Local\Temp\$LOGFILE
+    make test-e2e           | tee -a  C:\Users\Administrator.ANSIBLE-TEST-VS\AppData\Local\Temp\$LOGFILE
     Check-ExitCode $LASTEXITCODE
 
     Pop-Location
