+++ Makefile
@@ -164,122 +164,6 @@ test:
 test-windows:
 	go test $(UNIT_TEST_ARGS)  $(PKGS)
 
-.PHONY: test-generic
-test-generic: install ## Run generic integration tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo generic" tests/integration/
-
-.PHONY: test-cmd-login-logout
-test-cmd-login-logout: install ## Run odo login and logout tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS_SERIAL) -focus="odo login and logout command tests" tests/integration/loginlogout/
-
-.PHONY: test-cmd-link-unlink-4-cluster
-test-cmd-link-unlink-4-cluster: install ## Run link and unlink commnad tests against 4.x cluster
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo link and unlink commnad tests" tests/integration/
-
-.PHONY: test-cmd-project
-test-cmd-project: install ## Run odo project command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo project command tests" tests/integration/project/
-
-.PHONY: test-cmd-pref-config
-test-cmd-pref-config: install ## Run odo preference and config command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo preference and config command tests" tests/integration/
-
-.PHONY: test-plugin-handler
-test-plugin-handler: install ## Run odo plugin handler tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo plugin functionality" tests/integration/
-
-.PHONY: test-cmd-devfile-list
-test-cmd-devfile-list: install ## Run odo list devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo list with devfile" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-init
-test-cmd-devfile-init: install ## Run odo init devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile init command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-push
-test-cmd-devfile-push: install ## Run odo push devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile push command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-exec
-test-cmd-devfile-exec: install ## Run odo exec devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile exec command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-status
-test-cmd-devfile-status: install ## Run odo status devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile status command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-watch
-test-cmd-devfile-watch: install ## Run odo devfile watch command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile watch command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-app
-test-cmd-devfile-app: install ## Run odo devfile app command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile app command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-delete
-test-cmd-delete: install ## Run odo delete command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo delete command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-registry
-test-cmd-devfile-registry: install ## Run odo devfile registry command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile registry command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-test
-test-cmd-devfile-test: install ## Run odo devfile test command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile test command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-debug
-test-cmd-devfile-debug: install ## Run odo debug devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile debug command tests" tests/integration/devfile/
-	$(RUN_GINKGO) $(GINKGO_FLAGS_SERIAL) -focus="odo devfile debug command serial tests" tests/integration/devfile/debug/
-
-.PHONY: test-cmd-devfile-storage
-test-cmd-devfile-storage: install ## Run odo storage devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile storage command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-log
-test-cmd-devfile-log: install ## Run odo log devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile log command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-env
-test-cmd-devfile-env: install ## Run odo env devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile env command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-devfile-config
-test-cmd-devfile-config: install ## Run odo config devfile command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile config command tests" tests/integration/devfile/
-
-.PHONY: test-cmd-watch
-test-cmd-watch: install ## Run odo watch command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo watch command tests" tests/integration/
-
-.PHONY: test-cmd-debug
-test-cmd-debug: install ## Run odo debug command tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo debug command tests" tests/integration/
-	$(RUN_GINKGO) $(GINKGO_FLAGS_SERIAL) -focus="odo debug command serial tests" tests/integration/debug/
-
-# Service, link and login/logout command tests are not the part of this test run
-.PHONY: test-integration
-test-integration: install ## Run command's integration tests irrespective of service catalog status in the cluster.
-	$(RUN_GINKGO) $(GINKGO_FLAGS) tests/integration/
-
-.PHONY: test-interactive
-test-interactive: install ## Run integration interactive tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) tests/interactive/
-
-.PHONY: test-integration-devfile
-test-integration-devfile: install ## Run devfile integration tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) tests/integration/devfile/
-	$(RUN_GINKGO) $(GINKGO_FLAGS_SERIAL) tests/integration/devfile/debug/
-
-.PHONY: test-e2e-devfile
-test-e2e-devfile: install ## Run devfile e2e tests: odo devfile supported tests
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile supported tests" tests/e2escenarios/
-
-.PHONY: test-e2e-all
-test-e2e-all: install ## Run all e2e test scenarios
-	$(RUN_GINKGO) $(GINKGO_FLAGS) tests/e2escenarios/
-
 # run make cross before this!
 .PHONY: packages
 packages: ## create deb and rpm packages using fpm in ./dist/pkgs/
@@ -298,6 +182,10 @@ vendor-update: ## Update vendoring
 openshiftci-presubmit-unittests:
 	./scripts/openshiftci-presubmit-unittests.sh
 
-.PHONY: test-cmd-devfile-describe
-test-cmd-devfile-describe: install
-	$(RUN_GINKGO) $(GINKGO_FLAGS) -focus="odo devfile describe command tests" tests/integration/devfile/
+.PHONY: test-integration
+test-integration:
+	$(RUN_GINKGO) $(GINKGO_FLAGS) tests/integration
+
+.PHONY: test-e2e
+test-e2e:
+	$(RUN_GINKGO) $(GINKGO_FLAGS) tests/e2escenarios
