+++ scripts/openshiftci-periodic-tests.sh
@@ -28,15 +28,9 @@ oc whoami
 
 # Integration tests
 make test-integration || error=true
-make test-integration-devfile || error=true
-make test-cmd-login-logout || error=true
-make test-cmd-project || error=true
-
-# Interactive tests
-make test-interactive || error=true
 
 # E2e tests
-make test-e2e-all || error=true
+make test-e2e || error=true
 
 if [ $error ]; then
     exit -1
