+++ tests/integration/cmd_add_binding_test.go
@@ -1,4 +1,4 @@
-package devfile
+package integration
 
 import (
 	"fmt"
