+++ tests/integration/cmd_analyze_test.go
@@ -1,4 +1,4 @@
-package devfile
+package integration
 
 import (
 	"path/filepath"
