+++ tests/integration/cmd_describe_component_test.go
@@ -1,4 +1,4 @@
-package devfile
+package integration
 
 import (
 	"os"
