+++ tests/integration/cmd_describe_list_binding_test.go
@@ -1,4 +1,4 @@
-package devfile
+package integration
 
 import (
 	"fmt"
