+++ tests/integration/cmd_devfile_build_images_test.go
@@ -1,4 +1,4 @@
-package devfile
+package integration
 
 import (
 	"fmt"
