+++ tests/integration/cmd_devfile_deploy_test.go
@@ -1,4 +1,4 @@
-package devfile
+package integration
 
 import (
 	"fmt"
