--- tests/integration/devfile/debug/debug_suite_test.go
@@ -1,11 +1,0 @@
-package debug
-
-import (
-	"testing"
-
-	"github.com/redhat-developer/odo/tests/helper"
-)
-
-func TestDebug(t *testing.T) {
-	helper.RunTestSpecs(t, "Devfile Debug Suite")
-}
