--- tests/integration/devfile/devfile_suite_test.go
@@ -1,11 +1,0 @@
-package devfile
-
-import (
-	"testing"
-
-	"github.com/redhat-developer/odo/tests/helper"
-)
-
-func TestDevfiles(t *testing.T) {
-	helper.RunTestSpecs(t, "Devfile Suite")
-}
