+++ tests/integration/interactive_dev_test.go
@@ -1,7 +1,7 @@
 //go:build linux || darwin || dragonfly || solaris || openbsd || netbsd || freebsd
 // +build linux darwin dragonfly solaris openbsd netbsd freebsd
 
-package interactive
+package integration
 
 import (
 	"fmt"
