--- tests/integration/project/project_suite_test.go
@@ -1,11 +1,0 @@
-package project
-
-import (
-	"testing"
-
-	"github.com/redhat-developer/odo/tests/helper"
-)
-
-func TestProject(t *testing.T) {
-	helper.RunTestSpecs(t, "Project Suite")
-}
